+++ clients/unity/Corpc/ProtoMessage.cs
@@ -7,14 +7,14 @@ namespace Corpc
 {
     public class ProtoMessage
     {
-        public short Type { get; }
+        public int Type { get; }
         public ushort Tag { get; }
         public IMessage Data { get; }
 
         // NeedCrypter只用于发送时
         public bool NeedCrypter { get; }
 
-        public ProtoMessage(short type, ushort tag, IMessage data, bool needCrypter)
+        public ProtoMessage(int type, ushort tag, IMessage data, bool needCrypter)
         {
             Type = type;
             Tag = tag;
